+++ sync-for-expenses/src/sdk/models/operations/uploadexpenseattachment.ts
@@ -20,6 +20,9 @@ export class UploadExpenseAttachmentRequest extends SpeakeasyBase {
     })
     requestBody?: UploadExpenseAttachmentRequestBody;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
@@ -43,6 +46,9 @@ export class UploadExpenseAttachmentResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     attachment?: shared.Attachment;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -52,9 +58,15 @@ export class UploadExpenseAttachmentResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }